requirements.txt
aiohttp==3.8.3
aiosignal==1.3.1
async-timeout==4.0.2
attrs==22.2.0
Authlib==1.2.0
certifi==2022.12.7
cffi==1.15.1
charset-normalizer==2.1.1
click==8.1.3
cryptography==39.0.0
decorator==5.1.1
Flask==2.2.2
frozenlist==1.3.3
idna==3.4
importlib-metadata==6.0.0
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.2
multidict==6.0.4
numpy==1.24.1
openai==0.26.2
pandas==1.5.3
pycparser==2.21
python-dateutil==2.8.2
pytz==2022.7.1
requests==2.28.2
six==1.16.0
tqdm==4.64.1
urllib3==1.26.14
validators==0.19.0
weaviate-client==3.11.0
Werkzeug==2.2.2
yarl==1.8.2
zipp==3.11.0
